+++ CHANGELOG.md
@@ -7,7 +7,11 @@
 ### Breaking Changes
 
 ### Bug Fixes
-- Updating logic to check against the error first before checking the call object
+- Fix the order of CompositeError throwing in joining a call
+- Fix for info header doesn't disappear when bottom drawer is displayed
+- Fix for the local video rotation in setup view when device orientation is locked
+- Fix the display of the participant with empty name in the participant list. 
+- Fix the default audio selection when the UI Composite is launched. 
 
 ## 1.0.0-beta.1 (2021-12-09)
 This is the initial release of Azure Communication UI Library. For more information, please see the [README](README.md) and [QuickStart](https://docs.microsoft.com/en-us/azure/communication-services/quickstarts/ui-library/get-started-call?tabs=kotlin&pivots=platform-ios).