+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Project improvements
 - Update Redux State when initializing the local video preview in the setup view. [#23](https://github.com/Azure/communication-ui-library-ios/pull/23)
+- Remove alphabet in CFBundleShortVersionString. [#27](https://github.com/Azure/communication-ui-library-ios/pull/27)
 - Code cleanup and refactoring.